requirements.txt
transformers==4.30.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.23.5
nltk==3.8.1
pandas==2.0.3
tqdm==4.65.0